refactor: 点位删除和新增的逻辑优化，减少无效代码

+++ src/pages/pageMap/components/PopupContent.vue
@@ -1,9 +1,8 @@
 <script lang="ts" setup>
 import type L from 'leaflet'
-import { ElButton, ElMessage, ElMessageBox } from 'element-plus'
+import { ElButton, ElMessageBox } from 'element-plus'
 import { ceil } from 'lodash'
-import { useMarkerEdit, useMarkerStage } from '../hooks'
-import { Logger } from '@/utils'
+import { useMarkerDelete } from '../hooks'
 import { useUserStore } from '@/stores'
 
 const props = defineProps<{
@@ -12,81 +11,31 @@ const props = defineProps<{
 }>()
 
 const emits = defineEmits<{
-  (e: 'clickOutside', v: PointerEvent): void
   (e: 'command', v: string): void
   (e: 'refresh'): void
 }>()
 
-const logger = new Logger('[点位编辑]')
-
-const markerInfoRef = toRef(props, 'markerInfo')
-
 /** 用户信息 */
 const userStore = useUserStore()
 
-// 管理员直接走点位删除方法
-const { deleteMarker, onDeleteSuccess, onDeleteError } = useMarkerEdit(markerInfoRef)
-onDeleteSuccess(() => {
-  ElMessage.success('删除成功')
-  emits('refresh')
-})
+const markerInfoRef = toRef(props, 'markerInfo')
 
-onDeleteError((err) => {
-  ElMessage.error(err.message)
-})
-
-/** 初始化暂存点位信息 */
-const punctuateData: API.MarkerPunctuateVo = {
-  author: userStore.info.id,
-  punctuateId: props.markerInfo.id,
-  originalMarkerId: props.markerInfo.id,
-  methodType: 3, // 删除
-  status: 0,
-  version: props.markerInfo.version ? props.markerInfo.version + 1 : 1,
-  ...props.markerInfo,
-}
-
-/** 通用错误处理 */
-const commonErrorHandler = (err: Error) => ElMessage.error(err.message)
-
-// 非管理员走暂存审核方法
-const { stageMarker, onStageError, onPushStagedSuccess, onPushStagedError } = useMarkerStage(ref(punctuateData))
-onPushStagedSuccess(() => {
-  ElMessage.success('提交点位编辑请求成功')
-})
-onPushStagedError(commonErrorHandler)
-onStageError(commonErrorHandler)
+const { deleteMarker, onSuccess } = useMarkerDelete(markerInfoRef)
 
 /** 删除事件 */
 const onClickDel = async () => {
-  try {
-    // 打点员删除需要过审核
-    if (!userStore.isAdmin) {
-      await ElMessageBox.confirm('删除操作将被提交到审核，确认删除点位？', '确认操作', {
-        type: 'warning',
-      })
-      await stageMarker()
-      return
-    }
-    // admin直接请求
-    await ElMessageBox.confirm('该操作不可逆，确认删除点位？', '确认操作', {
-      type: 'warning',
-    })
-    await deleteMarker()
-  }
-  catch (err) {
-    logger.info(err)
-  }
+  const isConfirm = await ElMessageBox.confirm(
+    `${userStore.isAdmin ? '操作不可逆' : '删除操作将被提交到审核'}，确认删除点位？`,
+    '删除点位',
+    { type: 'warning' },
+  ).then(() => true).catch(() => false)
+  isConfirm && await deleteMarker()
 }
-
-const containerRef = ref<HTMLElement | null>(null)
-onClickOutside(containerRef, (ev) => {
-  emits('clickOutside', ev)
-})
+onSuccess(() => emits('refresh'))
 </script>
 
 <template>
-  <div ref="containerRef" class="w-full h-full flex flex-col gap-2">
+  <div class="w-full h-full flex flex-col gap-2">
     <div class="font-bold">
       {{ markerInfo.id }} - {{ markerInfo.markerTitle }}
     </div>
